requirements.txt
Django==2.2.7
django-filter==2.2.0
django-guardian==2.1.0
djangorestframework==3.10.3
djangorestframework-api-key==1.4.1
djangorestframework-simplejwt==4.3.0
Markdown==3.1.1
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
mysqlclient-1.4.4-cp37-cp37m-win32.whl ; sys_platform == "win32"
django-dbrouter